refactor(iocounter): move iocounter back as it is used by external packages
We should refactor those packages to not use `iocounter` as it probably
shouldn't be a publically accessible package, but other packages outside
of flux were using it.
