+++ iocounter/writer.go
